+++ dianjin-20240628/alibabacloud_dianjin20240628/models/_realtime_dialog_assist_request.py
@@ -17,7 +17,9 @@ class RealtimeDialogAssistRequest(DaraModel):
         hang_up_dialog: bool = None,
         meta_data: Dict[str, Any] = None,
         request_id: str = None,
+        script_content_played: str = None,
         session_id: str = None,
+        user_vad: bool = None,
     ):
         self.analysis = analysis
         # This parameter is required.
@@ -30,8 +32,10 @@ class RealtimeDialogAssistRequest(DaraModel):
         self.meta_data = meta_data
         # This parameter is required.
         self.request_id = request_id
+        self.script_content_played = script_content_played
         # This parameter is required.
         self.session_id = session_id
+        self.user_vad = user_vad
 
     def validate(self):
         if self.conversation_model:
@@ -67,8 +71,14 @@ class RealtimeDialogAssistRequest(DaraModel):
         if self.request_id is not None:
             result['requestId'] = self.request_id
 
+        if self.script_content_played is not None:
+            result['scriptContentPlayed'] = self.script_content_played
+
         if self.session_id is not None:
             result['sessionId'] = self.session_id
+
+        if self.user_vad is not None:
+            result['userVad'] = self.user_vad
 
         return result
 
@@ -98,26 +108,38 @@ class RealtimeDialogAssistRequest(DaraModel):
         if m.get('requestId') is not None:
             self.request_id = m.get('requestId')
 
+        if m.get('scriptContentPlayed') is not None:
+            self.script_content_played = m.get('scriptContentPlayed')
+
         if m.get('sessionId') is not None:
             self.session_id = m.get('sessionId')
+
+        if m.get('userVad') is not None:
+            self.user_vad = m.get('userVad')
 
         return self
 
 class RealtimeDialogAssistRequestConversationModel(DaraModel):
     def __init__(
         self,
+        begin: int = None,
+        begin_time: str = None,
         content: str = None,
         customer_id: str = None,
         customer_service_id: str = None,
         customer_service_type: str = None,
+        end: int = None,
         role: int = None,
         type: str = None,
     ):
+        self.begin = begin
+        self.begin_time = begin_time
         # This parameter is required.
         self.content = content
         self.customer_id = customer_id
         self.customer_service_id = customer_service_id
         self.customer_service_type = customer_service_type
+        self.end = end
         # This parameter is required.
         self.role = role
         self.type = type
@@ -130,6 +152,12 @@ class RealtimeDialogAssistRequestConversationModel(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.begin is not None:
+            result['begin'] = self.begin
+
+        if self.begin_time is not None:
+            result['beginTime'] = self.begin_time
+
         if self.content is not None:
             result['content'] = self.content
 
@@ -142,6 +170,9 @@ class RealtimeDialogAssistRequestConversationModel(DaraModel):
         if self.customer_service_type is not None:
             result['customerServiceType'] = self.customer_service_type
 
+        if self.end is not None:
+            result['end'] = self.end
+
         if self.role is not None:
             result['role'] = self.role
 
@@ -152,6 +183,12 @@ class RealtimeDialogAssistRequestConversationModel(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('begin') is not None:
+            self.begin = m.get('begin')
+
+        if m.get('beginTime') is not None:
+            self.begin_time = m.get('beginTime')
+
         if m.get('content') is not None:
             self.content = m.get('content')
 
@@ -163,6 +200,9 @@ class RealtimeDialogAssistRequestConversationModel(DaraModel):
 
         if m.get('customerServiceType') is not None:
             self.customer_service_type = m.get('customerServiceType')
+
+        if m.get('end') is not None:
+            self.end = m.get('end')
 
         if m.get('role') is not None:
             self.role = m.get('role')